refactor: extract i18n resolution from pdf/mapping to report_i18n
Move is_i18n_ref, human_source, and resolve_i18n from
adapters/pdf/mapping.py to report_i18n.py (their natural home).

This eliminates the shared→adapter boundary violation where
shared/run_context.py had a deferred import from adapters/pdf/mapping.
Also removes the now-redundant cast() in run_context.py.

Updates 4 importers (1 production, 1 shared, 2 test files).

Closes #702
